[fibsem] Give the shared view projection a view tilt and a neutral home
`view_corrected_stage_movement` and `inverse_view_corrected_dy` were written for
fluorescence and lived in `fibsem/fm/reprojection.py` with the camera tilt baked in.
They are not fluorescence-specific: they are the microscope-free form of
`FibsemMicroscope._view_corrected_stage_movement(expected_y, view_tilt)`, whose three
callers are the electron column (0), the ion column (its column tilt) and the camera
(its camera tilt). Only the last was hardcoded.

So they take a `view_tilt` argument and move to `fibsem/transformations.py`, next to
the other angle maths over stage tilt and pre-tilt. The fluorescence callers pass
`camera_tilt` and are otherwise unchanged; the beam side becomes a second caller
rather than a fourth copy of the formula.

Pure move plus the parameter, with no numerical change: forward, inverse and
round-trip dumped across 5 geometries x 8 tilts x 4 rotations x 5 displacements --
800 combinations, compared at `repr` rather than rounded, byte-identical before and
after. The 517 existing parity tests pass unchanged.
